Prepares for the new multiplayer conscious inventory system.
Major benefit here is that each instance of an item will refer to a shared data source, each trainer then does not have duplicate items with duplicate information. Making it better for multi-trainer use.

The new inventory system (next commit) will be independent of any one trainer, so it can be reusable. As well as better encapsulate inventory operations and functionality.